+++ src/Commands/Backup/RestoreCommand.php
@@ -4,6 +4,7 @@ namespace Pantheon\Terminus\Commands\Backup;
 
 use Pantheon\Terminus\Commands\WorkflowProcessingTrait;
 use Pantheon\Terminus\Exceptions\TerminusException;
+use Pantheon\Terminus\Exceptions\TerminusNotFoundException;
 
 /**
  * Class RestoreCommand
@@ -31,10 +32,33 @@ class RestoreCommand extends SingleBackupCommand
      */
     public function restoreBackup($site_env, array $options = ['file' => null, 'element' => 'all',])
     {
+        $this->validateElement($site_env, $options['element'], true);
+
         $site = $this->getSite($site_env);
         $env = $this->getEnv($site_env);
 
-        $backup = $this->getBackup($site_env, $options);
+        if ($options['file'] && $options['element'] !== 'all') {
+            throw new TerminusException('You cannot specify a file and an element at the same time.');
+        }
+
+        $elements = [];
+        if ($options['element'] === 'all') {
+            $elements = ['code', 'files', 'database'];
+        } else {
+            $elements[] = $this->getElement($options['element']);
+        }
+
+        $backups = $this->doGetBackups($site_env, $elements, $options);
+
+        if (!$backups) {
+            throw new TerminusNotFoundException(
+                'No backups available. Create one with `terminus backup:create {site}.{env}`',
+                [
+                    'site' => $this->getSite($site_env)->getName(),
+                    'env' => $env->getName(),
+                ]
+            );
+        }
 
         if (!$this->confirm(
             'Are you sure you want to restore to {env} on {site}?',
@@ -43,16 +67,55 @@ class RestoreCommand extends SingleBackupCommand
             return;
         }
 
-        $workflow = $backup->restore();
-        try {
-            $this->processWorkflow($workflow);
-            $this->log()->notice('Restored the backup to {env}.', ['env' => $env->id,]);
-        } catch (\Exception $e) {
-            $message = $workflow->getMessage();
-            if (trim($message) == 'Successfully queued restore_site') {
-                $message = 'There was an error while restoring your backup.';
+        $this->doRestoreBackups($backups, $env->id);
+    }
+
+    /**
+     * Get backups for given elements.
+     *
+     * @param string $site_env
+     *   Site & environment in the format `site-name.env`
+     * @param array $elements
+     *   Elements to get backups for.
+     * @param array $options
+     *   Options to pass to the getBackup function.
+     */
+    protected function doGetBackups(string $site_env, array $elements, array $options)
+    {
+        $backups = [];
+        foreach ($elements as $element) {
+            $options['element'] = $element;
+            try {
+                $backups[$element] = $this->getBackup($site_env, $options);
+            } catch (TerminusNotFoundException $e) {
+                continue;
             }
-            throw new TerminusException($message);
+        }
+        return $backups;
+    }
+
+    /**
+     * Restore backups to given site.
+     *
+     * @param array $backups
+     *   Backup files to restore
+     * @param string $env
+     *   Environment name.
+     */
+    protected function doRestoreBackups(array $backups, string $env)
+    {
+        foreach ($backups as $type => $backup) {
+            $workflow = $backup->restore();
+            try {
+                $this->processWorkflow($workflow);
+                $this->log()->notice('Restored the {type} backup to {env}.', ['env' => $env, 'type' => $type]);
+            } catch (\Exception $e) {
+                $message = $workflow->getMessage();
+                if (trim($message) == 'Successfully queued restore_site') {
+                    $message = 'There was an error while restoring your backup.';
+                }
+                throw new TerminusException($message);
+            }
         }
     }
 }