+++ tests/Functional/SiteCommandsTest.php
@@ -88,5 +88,12 @@ class SiteCommandsTest extends TerminusTestBase
         $this->assertIsArray($siteInfo);
         $this->assertArrayHasKey('organization', $siteInfo);
         $this->assertEquals($this->getOrg(), $siteInfo['organization']);
+        $this->assertArrayHasKey('framework', $siteInfo);
+        // Format json prints raw value.
+        $this->assertEquals('drupal8', $siteInfo['framework']);
+
+        // Format table prints processed value.
+        $siteInfoTable = $this->terminus(sprintf('site:info %s', $this->mockSiteName));
+        $this->assertStringContainsString('Drupal 8 or later', $siteInfoTable);
     }
 }